moved goify, gotyperef and gotypename to types.go
